Added the JavaFX library
There are some issue with the importation of the library.
If I run using the menu bar Run commands, the console will complain about JavaFX runtime components being missing, most of the time.
If I run use the Run | Debug options next to the main method, it works.
I have no idea why but I need to fix this for convenience's sake.